Update README.md
Completed an interesting Excel project (Project#1). In this project I was able to learn more about the Excel Pivot Table functions, Excel Dash-Board Design and some useful formulas/function (sum, sumif, count, countif, vlookup, hlookup etc.)  

In this project, I was assigned as Junior Data Analyst and also given  data to figure out the following questions:- 

-What is the most used ship mode by our customer ? 
- Who are our top 5 customer in terms of sales ? 
-which segment of client generates the most sales ?
-which city has the most sales value ?
-which state generates the most sales values?
-what are the top performing product categories in terms of sales & profit 
-what is the most profitable product that we sell?
-what is the lowest profitable product that we sell?
-.on average how long does it take order to reach our clients ? Based on the each shipping mode 

In this project not all I was able to learn pivot tables but it I was also able to learn and create eye catching dashboards. 

This project serves as a valuable learning opportunity for me.